内部采购单

+++ src/main/java/com/wondertek/cpm/web/rest/ContractBudgetResource.java
@@ -4,13 +4,17 @@ import com.codahale.metrics.annotation.Timed;
 import com.wondertek.cpm.domain.ContractBudget;
 import com.wondertek.cpm.domain.ContractInfo;
 import com.wondertek.cpm.domain.vo.ContractBudgetVo;
+import com.wondertek.cpm.domain.vo.LongValue;
+import com.wondertek.cpm.security.SecurityUtils;
 import com.wondertek.cpm.service.ContractBudgetService;
 import com.wondertek.cpm.service.ContractInfoService;
+import com.wondertek.cpm.service.ProjectInfoService;
 import com.wondertek.cpm.web.rest.util.HeaderUtil;
 import com.wondertek.cpm.web.rest.util.PaginationUtil;
 
 import io.swagger.annotations.ApiParam;
 
+import org.hibernate.validator.internal.util.Contracts;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -24,6 +28,7 @@ import javax.inject.Inject;
 
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -50,6 +55,8 @@ public class ContractBudgetResource {
     private ContractBudgetService contractBudgetService;
     @Inject
     private ContractInfoService contractInfoService;
+    @Inject
+    private ProjectInfoService projectInfoService;
 
     /**
      * POST  /contract-budgets : Create a new contractBudget.
@@ -58,18 +65,18 @@ public class ContractBudgetResource {
      * @return the ResponseEntity with status 201 (Created) and with body the new contractBudget, or with status 400 (Bad Request) if the contractBudget has already an ID
      * @throws URISyntaxException if the Location URI syntax is incorrect
      */
-    @PostMapping("/contract-budgets")
-    @Timed
-    public ResponseEntity<ContractBudget> createContractBudget(@RequestBody ContractBudget contractBudget) throws URISyntaxException {
-        log.debug("REST request to save ContractBudget : {}", contractBudget);
-        if (contractBudget.getId() != null) {
-            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("contractBudget", "idexists", "A new contractBudget cannot already have an ID")).body(null);
-        }
-        ContractBudget result = contractBudgetService.save(contractBudget);
-        return ResponseEntity.created(new URI("/api/contract-budgets/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert("contractBudget", result.getId().toString()))
-            .body(result);
-    }
+//    @PostMapping("/contract-budgets")
+//    @Timed
+//    public ResponseEntity<ContractBudget> createContractBudget(@RequestBody ContractBudget contractBudget) throws URISyntaxException {
+//        log.debug("REST request to save ContractBudget : {}", contractBudget);
+//        if (contractBudget.getId() != null) {
+//            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("contractBudget", "idexists", "A new contractBudget cannot already have an ID")).body(null);
+//        }
+//        ContractBudget result = contractBudgetService.save(contractBudget);
+//        return ResponseEntity.created(new URI("/api/contract-budgets/" + result.getId()))
+//            .headers(HeaderUtil.createEntityCreationAlert("contractBudget", result.getId().toString()))
+//            .body(result);
+//    }
 
     /**
      * PUT  /contract-budgets : Updates an existing contractBudget.
@@ -82,15 +89,52 @@ public class ContractBudgetResource {
      */
     @PutMapping("/contract-budgets")
     @Timed
-    public ResponseEntity<ContractBudget> updateContractBudget(@RequestBody ContractBudget contractBudget) throws URISyntaxException {
-        log.debug("REST request to update ContractBudget : {}", contractBudget);
-        if (contractBudget.getId() == null) {
-            return createContractBudget(contractBudget);
-        }
-        ContractBudget result = contractBudgetService.save(contractBudget);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert("contractBudget", contractBudget.getId().toString()))
-            .body(result);
+    public ResponseEntity<ContractBudgetVo> updateContractBudget(@RequestBody ContractBudgetVo contractBudgetVo) throws URISyntaxException {
+        log.debug("REST request to update ContractBudget : {}", contractBudgetVo);
+        ContractBudget contractBudget = new ContractBudget();
+        //是否构建项目
+        Boolean flag = contractBudgetService.checkByBudget(contractBudgetVo);
+        ContractBudget oldContractBudget = contractBudgetService.findOneById(contractBudgetVo.getId());
+        if (flag) {//构建项目
+        	contractBudget.setId(contractBudgetVo.getId());
+			contractBudget.setBudgetTotal(contractBudgetVo.getBudgetTotal());
+			contractBudget.setName(contractBudgetVo.getBudgetName());
+			contractBudget.setUpdateTime(ZonedDateTime.now());
+			contractBudget.setUpdator(SecurityUtils.getCurrentUserLogin());
+//			contractBudget.setContractId(oldContractBudget.getContractId());
+//			contractBudget.setDeptId(oldContractBudget.getDeptId());
+//			contractBudget.setUserName(contractBudget.getUserName());
+//			contractBudget.setUserId(oldContractBudget.getUserId());
+//			contractBudget.setStatus(1);
+//			contractBudget.setType(3);
+		}else {//未构建
+			contractBudget.setBudgetTotal(contractBudgetVo.getBudgetTotal());
+			contractBudget.setPurchaseType(contractBudgetVo.getPurchaseType());
+			contractBudget.setName(contractBudgetVo.getBudgetName());
+			contractBudget.setDept(contractBudgetVo.getDept());
+			contractBudget.setDeptId(contractBudgetVo.getDeptId());
+			contractBudget.setUserId(contractBudgetVo.getUserId());
+			contractBudget.setUserName(contractBudgetVo.getUserName());
+			contractBudget.setStatus(1);
+			contractBudget.setType(3);
+			//新建采购单
+			if (oldContractBudget == null) {
+				contractBudget.setContractId(contractBudgetVo.getContractId());
+				contractBudget.setUpdateTime(ZonedDateTime.now());
+				contractBudget.setCreateTime(ZonedDateTime.now());
+				contractBudget.setUpdator(SecurityUtils.getCurrentUserLogin());
+				contractBudget.setCreator(SecurityUtils.getCurrentUserLogin());
+			}else {//完成更新
+				contractBudget.setId(oldContractBudget.getId());
+				contractBudget.setContractId(oldContractBudget.getContractId());
+				contractBudget.setUpdateTime(ZonedDateTime.now());
+				contractBudget.setUpdator(SecurityUtils.getCurrentUserLogin());
+				contractBudget.setCreateTime(oldContractBudget.getCreateTime());
+				contractBudget.setCreator(oldContractBudget.getCreator());
+			}
+		}
+        contractBudgetService.save(contractBudget);
+        return new ResponseEntity<>(contractBudgetVo,HttpStatus.OK);
     }
 
     /**
@@ -125,11 +169,14 @@ public class ContractBudgetResource {
         ContractBudgetVo contractBudgetVo = new ContractBudgetVo();
         if (contractBudget != null) {
         	contractBudgetVo.setBudgetTotal(contractBudget.getBudgetTotal());
+        	contractBudgetVo.setBudgetName(contractBudget.getName());
         	contractBudgetVo.setId(contractBudget.getId());
         	contractBudgetVo.setDept(contractBudget.getDept());
+        	contractBudgetVo.setDeptId(contractBudget.getDeptId());
         	contractBudgetVo.setStatus(contractBudget.getStatus());
         	contractBudgetVo.setUserName(contractBudget.getUserName());
         	contractBudgetVo.setPurchaseType(contractBudget.getPurchaseType());
+        	contractBudgetVo.setUserId(contractBudget.getUserId());
 		}
         if (contractInfo != null) {
 			contractBudgetVo.setSerialNum(contractInfo.getSerialNum());
@@ -148,6 +195,10 @@ public class ContractBudgetResource {
     @Timed
     public ResponseEntity<Void> deleteContractBudget(@PathVariable Long id) {
         log.debug("REST request to delete ContractBudget : {}", id);
+        ContractBudget contractBudget = contractBudgetService.findOne(id);
+        if (contractBudget.getStatus() != 1) {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectInfo.save.haveDeleted", "")).body(null);
+		}
         contractBudgetService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("contractBudget", id.toString())).build();
     }
@@ -207,12 +258,21 @@ public class ContractBudgetResource {
     public ResponseEntity<List<ContractBudgetVo>> getAllContractBudgetsByParams(
     		@RequestParam(value = "name") String name,
     		@RequestParam(value = "serialNum") String serialNum,
+    		@RequestParam(value = "budgetName") String budgetName,
     		@ApiParam Pageable pageable)
     	throws URISyntaxException{
     	log.debug("REST request to get a page of ContractBudget");
-    	Page<ContractBudgetVo> page = contractBudgetService.searchPage(name,serialNum,pageable);
+    	Page<ContractBudgetVo> page = contractBudgetService.searchPage(name,serialNum,budgetName,pageable);
     	HttpHeaders headers = PaginationUtil.generateSearchPaginationHttpHeaders(name, page,"/api/contract-budgets");
     	return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
     
+    @GetMapping("/contract-budgets/queryUserContract")
+    @Timed
+    public ResponseEntity<List<LongValue>> queryUserContract() throws URISyntaxException{
+		log.debug("REST request to queryUserContrac");
+		List<LongValue> list = contractBudgetService.queryUserContract();
+        return new ResponseEntity<>(list, null, HttpStatus.OK);
+    	
+    }
 }
